+++ scripts/rollHandlers/swade/swade-base.js
@@ -65,11 +65,7 @@ export class RollHandlerBaseSwade extends RollHandler {
   async _toggleStatus(event, actor, actionId, tokenId) {
     const existsOnActor = actor.effects.find(
       e => e.getFlag("core", "statusId") == actionId);
-
-    const filter = (e) => e.id === actionId;
-    let data = CONFIG.statusEffects.find(filter);
-    if (!data) data = SWADE.statusEffects.find(filter);
-
+    const data = game.swade.util.getStatusEffectDataById(actionId);
     data["flags.core.statusId"] = actionId;
     await canvas.tokens.get(tokenId).toggleEffect(data, {active: !existsOnActor});
   }
@@ -112,7 +108,9 @@ export class RollHandlerBaseSwade extends RollHandler {
 
   /** @private */
   async _adjustAttributes(event, actor, macroType, actionId) {
-    let attribute = actor.system[macroType];
+    let attribute = (macroType === 'powerPoints')
+      ? actor.system[macroType].general
+      : actor.system[macroType];
 
     if (!attribute) return;
 
@@ -132,7 +130,9 @@ export class RollHandlerBaseSwade extends RollHandler {
 
     let update = { data: {} };
 
-    update.data[macroType] = { value: value };
+    update.data[macroType] = (macroType === 'powerPoints')
+      ? { general: { value: value } }
+      : { value: value };
 
     await actor.update(update);
   }