+++ scripts/rollHandlers/swade/swade-br2sw.js
@@ -39,6 +39,9 @@ export class RollHandlerBR2SWSwade extends RollHandler {
       case "attribute":
         this._rollAttribute(event, actor, actionId, tokenId);
         break;
+      case "runningDie":
+        actor.rollRunningDie();
+        break;
       case "skill":
         this._rollSkill(event, actor, actionId, tokenId);
         break;
@@ -92,11 +95,7 @@ export class RollHandlerBR2SWSwade extends RollHandler {
   async _toggleStatus(event, actor, actionId, tokenId) {
     const existsOnActor = actor.effects.find(
       e => e.getFlag("core", "statusId") == actionId);
-
-    const filter = (e) => e.id === actionId;
-    let data = CONFIG.statusEffects.find(filter);
-    if (!data) data = SWADE.statusEffects.find(filter);
-
+    const data = game.swade.util.getStatusEffectDataById(actionId);
     data["flags.core.statusId"] = actionId;
     await canvas.tokens.get(tokenId).toggleEffect(data, {active: !existsOnActor});
   }
@@ -181,7 +180,9 @@ export class RollHandlerBR2SWSwade extends RollHandler {
 
   /** @private */
   async _adjustAttributes(event, actor, macroType, actionId) {
-    let attribute = actor.system[macroType];
+    let attribute = (macroType === 'powerPoints') 
+      ? actor.system[macroType].general
+      : actor.system[macroType];
 
     if (!attribute) return;
 
@@ -201,7 +202,9 @@ export class RollHandlerBR2SWSwade extends RollHandler {
 
     let update = { data: {} };
 
-    update.data[macroType] = { value: value };
+    update.data[macroType] = (macroType === 'powerPoints')
+      ? { general: { value: value } }
+      : { value: value };
 
     await actor.update(update);
   }